@@ -1,5 +1,6 @@
 import logging
-from typing import Awaitable, Callable
+from collections.abc import Awaitable
+from typing import Callable
 
 import pytest
 from aiohttp import web
